Fix: - Remove state updates from timeout. It is not necessary as they are already inside an async function - Update tests to remove unnecesary timeout
