+++ __tests__/atomic-state-tests.tsx
@@ -13,12 +13,7 @@ it("Should increase clicks' atom value", async () => {
   })
 
   await waitFor(() => {
-    const tm = setTimeout(() => {
-      expect(CountDisplay.queryByText(/count is /)!.innerHTML).toBe(
-        "count is 1"
-      )
-      clearTimeout(tm)
-    }, 1000)
+    expect(CountDisplay.queryByText(/count is /)!.innerHTML).toBe("count is 1")
   })
 })
 
@@ -33,12 +28,9 @@ it("Should show the double of the count atom", async () => {
   })
 
   await waitFor(() => {
-    const tm = setTimeout(() => {
-      expect(CountDisplay.queryByText(/double is /)!.innerHTML).toBe(
-        "double is 4"
-      )
-      clearTimeout(tm)
-    }, 1000)
+    expect(CountDisplay.queryByText(/double is /)!.innerHTML).toBe(
+      "double is 4"
+    )
   })
 })
 
@@ -54,11 +46,8 @@ it("Should update an atom's value on user input", async () => {
   })
 
   await waitFor(() => {
-    const tm = setTimeout(() => {
-      expect(NameDisp.queryByText(/Username:/)!.innerHTML).toBe(
-        "Username: inuyasha"
-      )
-      clearTimeout(tm)
-    }, 1000)
+    expect(NameDisp.queryByText(/Username:/)!.innerHTML).toBe(
+      "Username: inuyasha"
+    )
   })
 })
