+++ lib/index.js
@@ -214,7 +214,7 @@ function useAtomCreate(init) {
     var _h = (0, react_1.useState)(false), runEffects = _h[0], setRunEffects = _h[1];
     var hydrated = (0, react_1.useRef)(false);
     var updateState = (0, react_1.useCallback)(function (v) { return __awaiter(_this, void 0, void 0, function () {
-        var willCancel, newValue, _a, hasChanded, notifyIfValueIsDefault, shouldNotifyOtherSubscribers, _i, _b, cleanupFunction, err_2, _c, effects_1, effect, cancelStateUpdate, err_3, tm_1;
+        var willCancel, newValue, _a, hasChanded, notifyIfValueIsDefault, shouldNotifyOtherSubscribers, _i, _b, cleanupFunction, err_2, _c, effects_1, effect, cancelStateUpdate, err_3;
         var _this = this;
         return __generator(this, function (_d) {
             switch (_d.label) {
@@ -318,23 +318,11 @@ function useAtomCreate(init) {
                 case 15:
                     if (!willCancel) {
                         defaultAtomsValues[init.name] = newValue;
-                        if (is18) {
-                            if (shouldNotifyOtherSubscribers) {
-                                notify(init.name, hookCall, newValue);
-                            }
-                            // Finally update state
-                            setState(newValue);
+                        if (shouldNotifyOtherSubscribers) {
+                            notify(init.name, hookCall, newValue);
                         }
-                        else {
-                            tm_1 = setTimeout(function () {
-                                if (shouldNotifyOtherSubscribers) {
-                                    notify(init.name, hookCall, newValue);
-                                }
-                                // Finally update state
-                                setState(newValue);
-                                clearTimeout(tm_1);
-                            }, 0);
-                        }
+                        // Finally update state
+                        setState(newValue);
                     }
                     return [7 /*endfinally*/];
                 case 16: return [7 /*endfinally*/];
@@ -408,13 +396,12 @@ function useAtomCreate(init) {
     }, [init.name]);
     (0, react_1.useEffect)(function () {
         var handler = function (e) { return __awaiter(_this, void 0, void 0, function () {
-            var tm_2;
             return __generator(this, function (_a) {
                 if (e.hookCall !== hookCall) {
-                    tm_2 = setTimeout(function () {
-                        setState(e.payload);
-                        clearTimeout(tm_2);
-                    }, 0);
+                    // const tm = setTimeout(() => {
+                    setState(e.payload);
+                    //   clearTimeout(tm)
+                    // }, 0)
                 }
                 return [2 /*return*/];
             });
@@ -563,7 +550,7 @@ function filter(init) {
         }, [initialValue]);
         function renderValue(e) {
             return __awaiter(this, void 0, void 0, function () {
-                var newValue_1, tm_3, err_4;
+                var newValue_1, tm_1, err_4;
                 return __generator(this, function (_a) {
                     switch (_a.label) {
                         case 0:
@@ -579,9 +566,9 @@ function filter(init) {
                                 setFilterValue(newValue_1);
                             }
                             else {
-                                tm_3 = setTimeout(function () {
+                                tm_1 = setTimeout(function () {
                                     setFilterValue(newValue_1);
-                                    clearTimeout(tm_3);
+                                    clearTimeout(tm_1);
                                 }, 0);
                             }
                             return [3 /*break*/, 4];
