+++ mocks/Tes1.tsx
@@ -9,10 +9,6 @@ export const RenderCount = () => {
 
 export const IncreaseButton = () => {
   const setAtomValue = useDispatch(clicks)
-  useEffect(() => {
-    // Reset count when mounting to prevent conflicts with other tests
-    setAtomValue(0)
-  }, [])
   return (
     <div>
       <button onClick={() => setAtomValue((value) => value + 1)}>
