+++ mocks/Test3.tsx
@@ -1,14 +1,6 @@
 import React, { useEffect } from "react"
-import { useDispatch, useFilter, filter } from "../"
-import { clicks } from "./atoms"
-
-const clicksFilter = filter<number>({
-  name: "clicksFilter",
-  get({ get }) {
-    const count = get(clicks)
-    return count * 2
-  },
-})
+import { useDispatch, useFilter } from "../"
+import { clicks, clicksFilter } from "./atoms"
 
 export const RenderCountDouble = () => {
   const constDoubleCount = useFilter(clicksFilter)
@@ -17,6 +9,12 @@ export const RenderCountDouble = () => {
 
 export const IncreaseButton2 = () => {
   const setAtomValue = useDispatch(clicks)
+
+  useEffect(() => {
+    // Reset count when mounting to prevent conflicts with other tests
+    setAtomValue(0)
+  }, [])
+  
   return (
     <div>
       <button onClick={() => setAtomValue((value) => value + 1)}>
