+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "atomic-state",
-  "version": "1.8.3",
+  "version": "1.8.4",
   "description": "State managment library for React",
   "main": "lib/index.js",
   "repository": {
